Better informative view when no repositories are present with Jenkinsfiles in Organization Folders, or no branches are available with Jenkinsfiles in multibranch projects.
