+++ src/main/resources/jenkins/branch/OrganizationFolderEmptyView/main.jelly
@@ -1,7 +1,7 @@
 <!--
  ~ The MIT License
  ~
- ~ Copyright (c) 2013, CloudBees, Inc.
+ ~ Copyright (c) 2016, CloudBees, Inc.
  ~
  ~ Permission is hereby granted, free of charge, to any person obtaining a copy
  ~ of this software and associated documentation files (the "Software"), to deal
@@ -21,15 +21,27 @@
  ~ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  ~ THE SOFTWARE.
  -->
-
 <?jelly escape-by-default='true'?>
-<st:compress xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:d="jelly:define" xmlns:l="/lib/layout" xmlns:t="/lib/hudson" xmlns:f="/lib/form" xmlns:i="jelly:fmt">
-  <l:layout title="${%Welcome}">
-    <j:set var="view" value="${it}"/>
-    <!-- expose view to the scripts we include from owner -->
-    <st:include page="sidepanel.jelly"/>
-    <l:main-panel>
-      <st:include page="main.jelly"/>
-    </l:main-panel>
-  </l:layout>
-</st:compress>
+<div xmlns:j="jelly:core" xmlns:st="jelly:stapler" xmlns:f="/lib/form">
+    <div style="margin: 20px 0; padding: 10px 30px 20px 30px; background: #f8f8f8; border-radius: 10px; max-width: 55em;">
+        <h1>This folder is empty</h1>
+    
+        <p>
+            There are no repositories found that contain buildable projects.
+            Organization Folders automatically build and manage repositories
+            that
+            contain recognizable projects.
+        </p>
+    
+        <j:forEach var="projectFactory" items="${it.owner.projectFactories}">
+            <st:include it="${projectFactory}" page="getting-started"
+                optional="true" />
+        </j:forEach>
+    
+        <p>
+            Please
+            <a href="./configure">configure Project Sources</a> to get started or
+            <a href="./computation">re-run the Folder Computation</a> if the repositories have since been configured.
+        </p>
+    </div>
+</div>
